% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEIR.R
\docType{data}
\name{aguapublica}
\alias{aguapublica}
\title{Agua Publica: Agua Publica a nivel de Departamento del Peru.}
\format{Base de datos con la informacion del agua publica de los departamentos del Peru.}
\usage{
data(aguapublica)
}
\description{
Agua Publica: Agua Publica a nivel de Departamento del Peru.
}
\keyword{datasets}
