% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEIR.R
\docType{data}
\name{alcantarillado}
\alias{alcantarillado}
\title{Alcantarillado: Alcantarillado a nivel de Departamento del Peru.}
\format{Base de datos con la informacion del alcantarillado de los departamentos del Peru.}
\usage{
data(alcantarillado)
}
\description{
Alcantarillado: Alcantarillado a nivel de Departamento del Peru.
}
\keyword{datasets}
