% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEIR.R
\docType{data}
\name{electrico}
\alias{electrico}
\title{Electricidad: Electricidad a nivel de Departamento del Peru.}
\format{Base de datos con la informacion de la electricidad de los departamentos del Peru.}
\usage{
data(electrico)
}
\description{
Electricidad: Electricidad a nivel de Departamento del Peru.
}
\keyword{datasets}
