% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEIR.R
\docType{package}
\name{INEIR}
\alias{INEIR}
\alias{INEIR-package}
\title{Trabaja con informacion estadistica del Peru. Creado por BEST}
\description{
Trabaja con informacion estadistica del Peru. Creado por BEST
}
