% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEIR.R
\docType{data}
\name{informalidad}
\alias{informalidad}
\title{Informalidad: Informalidad a nivel de Departamento del Peru.}
\format{Base de datos con la informacion de la informalidad de los departamentos del Peru.}
\usage{
data(informalidad)
}
\description{
Informalidad: Informalidad a nivel de Departamento del Peru.
}
\keyword{datasets}
