% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEIR.R
\docType{data}
\name{PBI2007}
\alias{PBI2007}
\title{PBI 2007: PBI del ano 2007 a nivel de Departamento del Peru.}
\format{Base de datos con la informacion del PBI del ano 2007 de los departamentos del Peru.}
\usage{
data(PBI2007)
}
\description{
PBI 2007: PBI del ano 2007 a nivel de Departamento del Peru.
}
\keyword{datasets}
