% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEIR.R
\docType{data}
\name{pobreza}
\alias{pobreza}
\title{Pobreza: Pobreza a nivel de Departamento del Peru.}
\format{Base de datos con la informacion de la pobreza de los departamentos del Peru.}
\usage{
data(pobreza)
}
\description{
Pobreza: Pobreza a nivel de Departamento del Peru.
}
\keyword{datasets}
